Relayer: remove send struct from api, enforce input size limits

+++ ethereum/contracts/interfaces/relayer/IRelayProvider.sol
@@ -13,7 +13,7 @@ interface IRelayProvider {
    */
   function quoteDeliveryOverhead(
     uint16 targetChainId
-  ) external view returns (uint256 deliveryOverhead);
+  ) external view returns (uint128 deliveryOverhead);
 
   /**
    * @notice This function should provide a fixed fee for 1 unit of gas on targetChainId.
@@ -22,7 +22,7 @@ interface IRelayProvider {
    *
    * @param targetChainId - the chain that should be quoted for.
    */
-  function quoteGasPrice(uint16 targetChainId) external view returns (uint256 gasPriceSource);
+  function quoteGasPrice(uint16 targetChainId) external view returns (uint88 gasPriceSource);
 
   /**
    * @notice This function should provide a quote in USD of the native asset price for all supported
@@ -32,7 +32,7 @@ interface IRelayProvider {
    *
    * @param chainId - the chain that should be quoted for.
    */
-  function quoteAssetPrice(uint16 chainId) external view returns (uint256 usdPrice);
+  function quoteAssetPrice(uint16 chainId) external view returns (uint64 usdPrice);
 
   /**
    * @notice When calculating the receiverValue of a delivery or performing a refund, a portion of
@@ -60,7 +60,7 @@ interface IRelayProvider {
    */
   function quoteMaximumBudget(
     uint16 targetChainId
-  ) external view returns (uint256 maximumTargetBudget);
+  ) external view returns (uint192 maximumTargetBudget);
 
   /**
    * @notice This function should return a payable address on this (source) chain where all awards
